commands/fsck: don't complain about missing objects for empty files
Right now, the empty file is its own pointer, but our pointer parser
doesn't consider this to be a pointer.  We'll fix that in a future
change, but for now, that when we parse these files, we don't fail if
the object representing the empty SHA-256 value is missing on disk,
since we don't need it to be present.