commands: don't try to upload or download zero-length pointers
Right now, our code to parse pointers doesn't handle empty files, which
are their own pointer.  However, in the future, it will learn to do so.
When we do this, we don't want to upload or download the object with
this OID, so let's special-case this code path to avoid attempting to
transfer these objects over the network.