commands: make sure we set up repository where appropriate
The smudge, clean, and filter-process commands all operate on the
contents of the repository, since they all expect to read or write files
into the LFS store in the repository.  As a result, let's call
setupRepository to verify that we are, in fact, in a repository and if
not, produce an error message.