Allow users to explicitly limit scan depth
In some scenarios git-lfs is unable to correctly deduce how to bound
the range of commits to look and when determining what to push. For
example, with Gerrit you push to refs/for/master in lieu of pushing to
refs/heads/master.  When this happens we will attempt to push
substantially more LFS objects than necessary resulting in poor
performance.

In lieu of using this fallback, if the user has a reasonable idea of
the underlying push depth we should allow them to limit the search
range.  Because we're unable to directly pass parameters via the
underlying git hooks we need to specify this in the command
environment.