lfs/gitscanner_refs.go: fix scanRefsToChan comment
We update the comment describing the scanRefsToChan() internal
function to match the one for the newer scanRefsToTree() function,
which was added in commit 608bc8d53efe0bfc789dae4b934e18af69ebd8e5
in PR #4525.  The older comment for the scanRefsToChan() function
has been out of sync with the function's actual signature and
usage since at least commit 3085dc32c7e52f1b2522a19a9259d1612b8fa4a5
in PR #1743, when the function was changed to invoke a provided
callback instead of return a channel of pointer object wrappers.