lock: allow locking multiple files
Since it's useful to be able to lock multiple files, let's add this
functionality.  We iterate over each path provided and attempt to lock
it.  If we're using JSON, we report all successful locks at the end;
otherwise, we report all successful locks as we process them.  All
errors are reported to standard error immediately, and we additionally
exit 2 on failure.

Note that this causes an incompatible change to the JSON output, since
we now always output an array of locks.