Makefile: use version suffix in go generate command
Currently, we lack the `/v2` suffix on our `go generate` command, which
results in us downloading and running the command from Git LFS v1.5.6,
which is probably not what we intended.  In addition, Go “helpfully”
modifies our go.mod and go.sum to include new entries, which is also
undesirable.

Let's update the Makefile to invoke the proper package to avoid these
problems.