unlock: allow locking multiple files
Since it's useful to be able to unlock multiple files, let's add this
functionality.  We iterate over each path provided and attempt to unlock
it.  If we're using JSON, we report all successful unlocking attempts at
the end; otherwise, we report all successful unlocking attempts as we
process them.  All errors are reported to standard error immediately,
and we additionally exit 2 on failure.  As a result of these changes,
avoid exiting early in various cases and instead return an error.

Note that we don't currently accept multiple --id arguments, although we
could in the future.

Note that this causes an incompatible change to the JSON output, since
we now always output an array of locks.