urlconfig: anchor regexp for key matching
Currently, we don't anchor the regexp for matching keys.  That means
that if we have a value of http.URL.sslCert and
http.URL.sslCertPasswordProtected, the latter can be read when when
looking up the former.  To avoid that, let's anchor the regexp properly
against the key at the beginning and end.  In addition, let's escape the
dots to prevent matching the wrong string.