githistory/rewriter: Add path to rewrite Tree cache key
Duplicated files (same name & content) were pulled from the rewriteTree cache.
In this case, .gitattribute files was not appended but files were
changed to a pointer anyway.

With this, duplicated files are not pulled from cache (as long as their
full path is different) and .gitattribute patterns match files changed
to a pointer.

Fixes git-lfs/git-lfs#4628