fs: specify a file as existing if it's empty
If the object has a size of zero, then we already have its object: it's
the empty file, and we shouldn't need to download it from anywhere.  If
we inquire if it exists, then say it does, and say its location is the
system's equivalent of /dev/null.

The only tricky case is if we're de-duplicating, linking, or writing to
a file, in which case we should not try to use a link or replace on the
file, since we neither want to link nor replace the system's /dev/null.