tr: add a tool to build locales into the binary
Currently, Git LFS has the benefit of being a single, relocatable
binary.  While it is customary to store locales in /usr/share/locales on
Unix systems, this isn't necessarily the case on Windows, and we'd like
to preserve the ability of users to move the binary where they like.

As a result, let's add a program which can be run by go generate that
embeds these translations into the binary.  They are Base64-encoded to
make handling them a little easier and avoiding need to write giant
strings of backslashed escape sequences.  The reader will note that this
is based off the related command to embed the manual pages.