Makefile: build translations when compiling
Let's make sure we build the translations into the binary when possible.
Specifically, if the msgfmt command exists, let's compile the .po files
into .mo files and then embed their contents into the binary using the
generator command to do that.

Let's also ignore the built .mo files and the generated .pot template
files so we don't check them in.