Introduce a single translation
It will take some time for us to get the codebase translated.  However,
as a test of our codepaths and as a demonstration of what to do, let's
introduce a single translation, plus a simple Spanish translation that
translates only that string.  This allows an invocation like the
following to print a simple translated message:

  LC_ALL=es_ES.UTF-8 git lfs filter-process