vendor,go.*: update x/crypto and dependencies
Update the vendored golang.org/x/crypto module to the
latest version, which in turn updates the vendored copy
of the x/sys module.

While the portions of x/crypto vendored into this project do
not actually change at all, updating the module's entry in
vendor/modules.txt and go.{mod,sum} means we will not be
flagged by security scanners regarding CVE-2020-29652 in the
x/crypto/ssh library code.

The Git LFS client should not be affected by that security
issue, since we do not use the ssh component of the x/crypto
module.

The specific commands run to perform this update were:

  go get golang.org/x/crypto@latest &&
  go mod tidy && go mod vendor