t/t-ssh.sh: also perform test without ssh variant
It is currently possible to cause the "ssh with proxy
command in lfs.url" test to fail by setting the
GIT_SSH_VARIANT environment variable to something like
"simple".  This causes the error message from the
lfs-ssh-echo test helper program to contain a different
string, without the "--" end-of-options delimiter and
without a leading "-" character on the invalid argument,
because those are not included by Git LFS.

We therefore want to run two versions of this test, one
with GIT_SSH_VARIANT forcibly unset, to test the default
case, and one with that variable set but with a different
check on the error message from lfs-ssh-echo.

Also, in a previous commit we restored some of the argument
parsing logic in the lfs-ssh-echo test helper program,
so we now want to restore the check that the helper's error
message contains "git@127.0.0.1", not just a "-p" flag.