docs/man: do not support parentheses in headers
This reverts commit 08cb542b1d83ca611e2403150d26194f08e4d86d
from PR #4458 because that change, while it does resolve some
formatting issues in the output from the "git lfs help" command,
does not address the fact that ronn(1) is unable to process
parentheses in header text and therefore still generates
pages (particularly for the git-lfs-migrate(1) page) with
broken references for such headers.

Since we will rewrite the relevant manual pages in a subsequent
commit to remove the parentheses, we simply revert this change first.