commands: handle checkout --to with no args
At present if the "git lfs checkout" command is invoked with
the --to option, a file path, and a "stage" option (--base,
--ours, or --theirs) , but without any object path, a Go panic
is reported, because there is an attempt to reference the
non-existent first element of the "args" slice.  For example,
"git lfs checkout --to foo --base" reports a panic.

We address this issue by exiting with an error message if the
number of additional arguments is not exactly one when --to is
supplied along with accompanying file path and one of the
"stage" options.

This also covers the situation where the user supplies these
options plus spurious additional arguments beyond a single Git
LFS object path, since only a single object can be checked out
at a given stage into the file specified with --to.  (Note that
this object path is not treated as a filepath pattern, unlike
the normal "git lfs checkout" mode of operation.)