docs/man: update checkout man page
In a previous commit we revised the "git lfs checkout" command to
handle certain improper argument error conditions more gracefully,
including when too many paths to conficting files are provided along
with the --to option (plus the necessary file for the --to option
and one of the "stage" options: --base, --ours, or --theirs).

We therefore also update the git-lfs-checkout(1) manual page to
clarify that only a single path should be provided in this case,
and that it must reference a conflicting Git LFS object file,
not just any file with a merge conflict.

We also add some missing formatting in the Synposis section to
ensure the two alternative invocation formats are clearly
separated, and we adjust the terminology of some of the
user-provided arguments to align better with Git's manual and
with other Git LFS manual pages.  Specifically, we use <file>
to indicate any file in the filesystem, <glob-pattern> to indicate
a Git LFS glob pathspec pattern matched according to gitignore(5)
rules, and <path> (in this case, <conflict-obj-path>) to indicate
a path within the repository.  And we expand on the details of
each of these terms in the text of the manual page as well.

Finally, we refine and expand the Examples section by adding
an example use case for the --to option, and making all the
examples follow the same formatting style.