Add support for a fake locale for testing
Sometimes we'd like to test to see how much of or whether our code is
localized.  However, it's very burdensome to create a full localization
in order to do so.  To help us, let's create a special, autogenerated
locale, i-reverse, which is generated by simply reversing all the
English words.  Add a script to do this automatically for us.

However, don't generate this locale on typical builds, since users might
not have xgotext installed, and adding this locale would bloat the
binary needlessly in most cases.