.github/workflows: switch to actions/checkout@v2
The actions/checkout workflow looks like it may have had a
security-sensitive change to it in the latest v2.  It doesn't look like
it affects us, but let's update to be safe.

Switch to actions/checkout@v2 everywhere and use the fetch-depth: 0
argument, which makes us fetch the entire history.  This is the main
reason we stuck with v1 for so long, and it's required for certain parts
of our build and release process to function correctly.