t/fixtures/migrate.sh: align test repo names
We revise the names used for several test repositories
created by the "git lfs migrate" tests, aligning them with
their actual contents and with the majority of the other
names used by the test suite.

However, we do have to keep them short enough that we
avoid running afoul of the Windows PATH_LEN maximum of 260
characters, including the full path to our CI runner
environments.