tq: simplify HTTP 403 status code messages
Each of the upload transfer queue adapters reports an error
with a similar message in the case of a 403 response, and all
of these messages begin with an "http:" prefix (even the
one in the SSH adapter).  However, other status codes are not
reported the same way.

This prefix should either be removed from the translatable
message text, or removed entirely; we choose the latter option
here as the simplest.  We also now interpolate the status code
rather than make it a fixed part of the translatable string,
and start the messages with a capital letter.