commands/command_fetch.go: separate message prefix
Akin to several other commands such as "git lfs migrate" and
"git lfs prune", the "git lfs fetch" command prefixes a number
of its information messages with the name of the command, i.e.,
with the "fetch:" prefix.

To align with those other commands, we therefore remove this
command-specific prefix from our translatable message strings
and make it a fixed prefix for all languages.