commands/command_dedup.go,t: consistent messages
We convert several output messages to use a spelling of
 "de-duplication" consistent with all other messages from
the "git lfs dedup" command.

We also revise one message's text to clarify that it refers
to a Git LFS object file.

Note that one of these messages is not yet passed as a
translation string, but we will address that issue in a
subsequent commit.