commands,t: use Git LFS project name in messages
We convert a few messages output by the various commands to use
the full "Git LFS" project name instead of the shorter "LFS" variant,
which will make them consistent with how the project name appears
in other messages elsewhere.

Note that we leave the status messages in the transfer queue
using the shorter variant as they are intended to be brief, are
checked extensively in the test suite, and are familiar to users
at this point.