commands/command_ext.go: translate ext header only
In commit 1ddc2f2662b71d365ca7934c39b4ae56a60b5673 of
PR #4781 text translation support was added for the
"git lfs ext" command and a multi-line message was
concatenated.  However, the latter part of this text should
be fixed for all users, since it displays a listing of
the extension's settings values using defined terms.

We therefore separate the first line of the message and
let that be translated, as it contains the language-specific
header text.