tq,t: remove transfer queue message prefix
One error message output by the transfer queue currently
begins with a "tq:" prefix, which is unlike any other error
messages from that package, or from anywhere else, for that
matter.

We can therefore simplify our message string slightly by
simply removing this prefix from it.

Note that this message is not yet passed as a translation
string, but we will address that issue in a subsequent commit.