commands/lockverifier.go: restore message indent
In commit c744476cb73c8489699331fbfc93393786f5a6ed of
PR #4781 first part of a multi-line message was made
translatable, while the second part, which consists only of
an example "git config" command, was not.

That change is reasonable; however, some indentation
whitespace preceding the example command was lost, so we
restore it here.