commands,t: use capitalized Git in messages
We convert a few messages output by the various commands to use
the capitalized "Git" project name instead of the lowercase variant,
which will make them consistent with how the Git project name appears
in other messages elsewhere.

Note that some of these messages are not yet passed as translation
strings, but we will address that issue in a subsequent commit.