t/t-fsck.sh: add invalid objects test
Although we have a number of tests of the "git lfs fsck --pointers"
option added in PR #4525, we do not yet have any of the companion
"git lfs fsck --objects" option.  We therefore add a pair of tests,
one of checking for expected output when Git LFS object files are
corrupt or missing, and one checking for no output when no Git LFS
objects exist in the repository.

Note that at present we expect an exit code of 2, not 1, when the
"git lfs fsck --objects" command is unable to move a completely
missing Git LFS object file to the .git/lfs/bad directory.  We
expect to revise this behaviour and the relevant exit code checks
in a subsequent commit.