commands/command_fsck.go: reorder ref arguments
To be more consistent with the calling convention of many
methods in the "lfs" package, such as ScanRefs(), we reorder
the "include" and "exclude" arguments passed to several
internal functions of the "git lfs fsck" command to be
in that order specifically.

Note that we still pass these arguments in the opposite
order to the ScanRefRange() and ScanRefRangeByTree() methods,
but we will address those issues in subsequent commits.