commands,lfs: reverse ScanRefRangeByTree ref args
As noted in a previous commit, the ScanRefRangeByTree()
method currently reverses its "left" and "right" arguments
before passing them to scanRefsByTree().  Since, for historical
reasons, in the context of these GitScanner methods, the "left"
ref is interpreted as specifying all the commits to include,
while the "right" ref specifies all the ones to exclude, we
should maintain this convention here as well when passing these
arguments to the scanRefsByTree() function, where they are
called "include" and "exclude", respectively.

We therefore also reverse the arguments in the sole caller of
ScanRefRangeByTree(), in the "git lfs fsck" command.