lfs/gitscanner*.go: update scan function comments
In a prior commit we renamed the "left" and "right" ref
arguments for many GitScanner methods and internal functions
to be "include" and "exclude" instead.  We therefore now also
update the descriptive comments for these methods and a number
of others, and rephrase some of these to be more explicit and
complete.

Also, in commit bbc3e24767920f8a1ccffe7e6cc999980ae48765 of PR #3755
the ScanLeftToRemote() method was renamed as ScanRangeToRemote(),
so we adjust one comment which still referenced the old name.