commands: rename some left/right ref terms
Following from work in prior commits, we can rename several
variables and arguments to use "include" and "exclude" instead
of the terms "left" and "right", which are more ambiguous.
And in the case of the "git lfs push" command, we can just use
a generic "ref" variable name.

We also rename the internal function UploadLeftOrAll() to
UploadRangeOrAll(), which is somewhat more precise.