Revert "Merge pull request #4795 from bk2204/actions-checkout-v2"
actions/checkout@v2 "helpfully" overwrites the tag with the data from
the ref, erasing the annotated tag, and therefore breaking git describe,
which by default only checks annotated tags.  Note that every tag except
the one for the current head is preserved, so git describe uses an
annotation that is based off the most recent tag.  This behavior is
described in actions/checkout#290.

The original reporter claiming that a security fix had been applied has
not provided details, and the behavior described should not be a
vulnerability in a single-tenant Actions VM.  Therefore, revert to
actions/checkout@v1 to preserve functionality at the expense of a
behavior which does not appear to describe an actual vulnerability.

This reverts commit e3893b1402349b9481ccd8e1cd876d9e355cf848, reversing
changes made to eb0dc945e92a9a17116a5c2fdde572991ac12ac2.