lfs: add old hook content to the list of old hooks
Our hook code is designed to automatically replace old hooks that it
knows about with new ones.  However, in ed3decf7 ("use backticks around
commands in messages", 2022-01-26), we made a change to the hook and
accidentally forgot to include the old hook content.

Let's fix this by providing our old hook content, which contains a
template, and processing each of the corresponding old hooks to replace
the template.  This will cause unmodified hooks to be automatically
updated, just like we expect.