Makefile,debian,rpm: build man pages into dirs
We currently build our manual pages and the HTML versions of
them into a single "man" directory, which results in release
package archives on GitHub that contain all of these files
interleaved with each other in that one directory.

When we build Linux RPM and Debian packages, though, the man
pages are installed into their conventional locations under
/usr/share/man/man{1,5}.  (The HTML versions are not included
at all.)

For those downloading the release package archives and not using
the RPM or Debian packages, we can improve their experience by
placing the roff man pages into "man1" and "man5" subdirectories
corresponding to their respective sections of the manual, and the
HTML versions of the pages into an "html" subdirectory.