script/windows-installer: check Git during setup
In commit a5b751f541abb52e30b60aebe530f9dbd72838df we added
a GitFoundInPath() function which we check before executing
a full installation; if it returns False, we do not install
the Git LFS binary.

However, we still install the uninstaller and add the
installation path to the PATH and GIT_LFS_PATH environment
variables, which leads to a partial and incomplete setup;
also, the Inno Setup installer still displays a final dialog
box with the message "Setup has finished installing Git LFS
on your computer", which is confusing.

However, if we call our GitFoundInPath() check function in
the InitializeSetup() event function and return its value,
then if it returns False the entire setup process is halted
and no actual installation occurs at all.  This leads to a
cleaner and more intuitive experience for users.

This also aligns the installation and uninstallation steps,
since the uninstaller also checks GitFoundInPath() in its
initialization event function and if the return value is
False then it terminates immediately without making any
changes.