script/windows-installer: show deregistration text
When uninstalling Git LFS on Windows, as of commit
cbd52be073a340808253208ee4430b6f6451c271 we now use the
term "deregister" in part of the message we display when
a Git exectuable is found in an unconventional location.

Since we also use the term "registered" in another part
of that message, we change it to switch to the term
"deregistered" in the uninstallation case.