t/testhelpers.sh: refine clone helper functions
The clone_repo*() family of helper functions defined in our
t/testhelpers.sh shell library are used extensively throughout our
test suite to perform "git clone" commands and log their output for
verification.  After running the "git clone" command, these functions
change the current working directory to the cloned repository's working
tree and then write the log of the command's output into that directory.
Before creating the log file, however, they call the "git config"
command to register our git-credential-lfstest test utility as the
program Git should use to store and retrieve credentials.

One consequence of the way the majority of these functions operate is
that if the "git clone" command should fail, then because we run all
our tests with the "set -e" shell option, no log file is generated.
Instead, the functions return immediately and do not change the
working directory or write a log file.

The clone_repo_clientcert() function is an exception to this
pattern, as a result of the changes introduced to it in commit
9a00eb3a684b61d33318bdfb89e3a1820b064c57 of PR #1893, which were made
to work around a problem caused by the version of libcurl installed
with macOS version 10.12 ("Sierra").

In the clone_repo_clientcert() function, the "set +e" shell option was
temporarily set before running the "git clone" command, and command's
exit code captured before setting the "set -e" option again.  This
allowed the function to check the command's output for the macOS error
message specific to the libcurl problem, and if it was present, add a
flag string to the log file and not return an error code.  The calling
test could then simply skip the remainder of its normal actions if it
detected the flag string in the log file.

In a prior commit in this PR we reversed some of these changes because
the underlying issue is fixed in all recent macOS and libcurl releases,
and so we no longer need to try to detect specific exceptions in the
output of the clone operation in the clone_repo_clientcert() function.

We did not reverse the use of the "set +e" shell option, though, because
there is some value in trying to avoid an immediate test failure if
the "git clone" command does not succeed, as will occur with all the
other clone_repo*() functions.

Under most circumstances the behaviour of the other functions is
reasonable, as any error messages from the "git clone" commands will
still appear in the stdout and stderr streams captured by our CI job
test framework, and so may be analyzed there.

However, this leaves no way for tests using these functions to
programmatically check the output of a failed "git clone" attempt.
For instance, if they are looking for a specific string to indicate
that the clone was unable to find a valid Git executable and ran
an invalid one instead (as per CVE-2022-24826), we may want a test
to be able to search the log file from a clone operation even after
it fails.

Therefore, in order to ensure that clone operation log traces are saved
in all cases, we change all of the clone_repo*() functions to pipe the
output of the "git clone" command into a tee(1) command, which then
writes the log file.  The advantage of the "tee" command over a simple
redirection of stdout is that when the "set -e" option is set, the shell
only considers the exit code of the final command in a pipeline to
determine the success of the pipeline, so a non-zero exit code from the
"git clone" command will not cause the function to return immediately.

We then retrieve the "git clone" command's exit code from the shell's
PIPESTATUS[0] variable and return if the code is non-zero.  If the clone
operation was successful, though, we proceed to change the working
directory and move the log file into the new repository's working tree,
where our tests expect to find it, and then return after calling
the "git config" command.

In the clone_repo_clientcert() function specifically, this change now
allows us to remove the use of the "set +e" shell option, as we are
able to read the exit code from the "git clone" command without it.
We also no longer need separate statements to write the log file under
both success and failure conditions.